add prority,but the tasks can not switch correctly

+++ OS/types.h
@@ -4,7 +4,7 @@
 
 typedef signed char             int8;
 typedef signed short            int16;
-typedef signed long             int32;
+typedef signed int              int32;
 
 typedef signed char const       cint8;
 typedef signed short const      cint16;
@@ -12,10 +12,10 @@ typedef signed long const       cint32;
 
 typedef unsigned char           uint8;
 typedef unsigned short          uint16;
-typedef unsigned long           uint32;
+typedef unsigned int            uint32;
 
 typedef unsigned char const     ucint8;
 typedef unsigned short const    ucint16;
-typedef unsigned long const     ucint32;
+typedef unsigned int const      ucint32;
 
 #endif
